asterisk_integration/requirements.txt
rapidpro-python
python-dotenv
